refactor(tests): move tests from /bin to /lib
They are actually testing entities from /lib.
