+++ lib/stencil-bundle.spec.js
@@ -13,8 +13,8 @@ const describe = lab.describe;
 const themePath = Path.join(process.cwd(), 'test/_mocks/themes/valid');
 const expect = Code.expect;
 const it = lab.it;
-const StencilBundle = require('../lib/stencil-bundle');
-const LangAssembler = require('../lib/lang-assembler');
+const StencilBundle = require('./stencil-bundle');
+const LangAssembler = require('./lang-assembler');
 const jspm = require('jspm');
 const themeSchema = Fs.readFileSync((Path.join(themePath, 'schema.json'))).toString();
