+++ lib/stencil-init.spec.js
@@ -9,9 +9,9 @@ const describe = lab.describe;
 const Inquirer = require('inquirer');
 const expect = Code.expect;
 const it = lab.it;
-const StencilInit = require('../lib/stencil-init');
+const StencilInit = require('./stencil-init');
 const JspmAssembler = require('../lib/jspm-assembler');
-const ThemeConfig = require('../lib/theme-config');
+const ThemeConfig = require('./theme-config');
 
 describe('stencil init', () => {
     let sandbox;
